align test framework to junit and renmae test class
